requirements.txt
alabaster==0.7.12
Babel==2.9.0
bdp==0.2.1
build==0.4.0
certifi==2020.6.20
chardet==4.0.0
cycler==0.10.0
data==0.4
decorator==4.4.2
docutils==0.16
funcsigs==1.0.2
future==0.18.2
idna==2.10
imagesize==1.2.0
importlib-metadata==4.0.1
Jinja2==2.11.3
kiwisolver==1.2.0
latex==0.7.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mpmath==1.1.0
numpy==1.19.2
opencv-python==4.5.2.54
packaging==20.9
pandas==1.1.5
pep517==0.10.0
pexpect==4.8.0
Pillow==7.2.0
pkg-resources==0.0.0
plotly==4.14.3
ptyprocess==0.7.0
Pygments==2.8.1
pygraphviz==1.6
pyparsing==2.4.7
pypi-install==0.0.5
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
retrying==1.3.3
SchemDraw==0.6.0
scipy==1.5.2
seaborn==0.11.1
shutilwhich==1.1.0
six==1.15.0
snowballstemmer==2.1.0
Sphinx==3.5.2
sphinx-argparse==0.2.5
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sympy==1.6.2
tempdir==0.7.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.3
zipp==3.4.1
